# By k_lyashkevich (173) and others
# Via k_lyashkevich (60) and others
* commit '4ab044366f33b7b402c1801c8405869ff4fb1b14 [formerly 320a1769b54e05414d9ac83a2888a0f8372c5c90]': (587 commits)
  [DF-9935] Code format only
  [DF-9935] Updated Win32 RP binary
  [DF-9935] Updated OSX RP binary
  [DF-9935] Updated asset cache key for packer
  [DF-9907] details by review re-fixed
  [DF-9927, DF-9928] Additional check on compression format set.
  [DF-9907] code fixed by code review, bug fixed
  [DF-9582] review fix
  [DF-9907] move fixed
  [DF-9912] restoring missed connection
  [DF-9163] Win64 memory profiler "fixed"
  [DF-9163] Code formatted
  [DF-9163] ResourcePacker rebuilt for OS X
  [DF-9163] ResourcePacker rebuilt
  [DF-9582] ugly bug fixed
  [DF-9870] Fixed ruler scale
  [DF-9582] clear tmp variables
  [DF-9582] selection mechanism fixed
  DF-9841 cmake dylibrefactoring
  [DF-9582] super hell bug fixed
  ...


Former-commit-id: c10a4650fb4f8b33482134ec7e64d465fe6c5734

+++ Sources/Internal/Collision/Collisions.cpp
@@ -155,7 +155,7 @@ bool Collisions::Is2DPolygonIntersects2DPolygon(const Polygon2& poly1,
                 float fpntY = poly1.points[pnt].y;
                 float fiY = poly2.points[i].y;
                 float fjY = poly2.points[j].y;
-                float t = (float)(fpntY - fiY) / (float)(fjY - fiY);
+                float t = (fpntY - fiY) / (fjY - fiY);
                 if (poly2.points[i].x + t * (poly2.points[j].x - poly2.points[i].x) >= poly1.points[pnt].x)
                 {
                     count++;
